@@ -1,11 +1,11 @@
-package com.chii.tt2info.ui;
+package com.chii.tt2info.ui.activity;
 
 import android.app.Activity;
 import android.content.Intent;
 import android.os.Bundle;
-import android.support.annotation.Nullable;
-import android.support.v7.app.ActionBar;
-import android.support.v7.app.AppCompatActivity;
+import androidx.annotation.Nullable;
+import androidx.appcompat.app.ActionBar;
+import androidx.appcompat.app.AppCompatActivity;
 import android.text.Editable;
 import android.text.TextWatcher;
 import android.util.Log;
@@ -29,7 +29,7 @@ import java.util.HashMap;
 import butterknife.BindView;
 import butterknife.ButterKnife;
 
-import static com.chii.tt2info.ui.MainActivity.REQUESTCODE_FROM_REGISTER;
+import static com.chii.tt2info.ui.activity.MainActivity.REQUESTCODE_FROM_REGISTER;
 import static com.chii.tt2info.connes.MyVolley.signin_url;
 
 public class LoginActivity extends AppCompatActivity {
